Import using correct service during a service exception day
